backend/requirements.txt
django
djagorestframework
environs